fanotify: Document "pre content" events
The new events FAN_PRE_ACCESS, FAN_PRE_MODIFY are created before access
to file or directory, to provides an opportunity for the event listener
to modify the content of the object before the user can accesss it.

Those events are available for group in class FAN_CLASS_PRE_CONTENT
They are reported with FAN_EVENT_INFO_TYPE_RANGE info record.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>